fix: native layer - check if exists before delete files

+++ android/app/src/main/java/org/ZingoLabs/Zingo/RPCModule.kt
@@ -39,6 +39,18 @@ class RPCModule internal constructor(private val reactContext: ReactApplicationC
         }
     }
 
+    fun walletBackup_exists(): Boolean {
+        // Check if a wallet already exists
+        val file = getFile(walletBackupFileName)
+        return if (file.exists()) {
+            Log.i("SCHEDULED_TASK_RUN", "Wallet backup exists")
+            true
+        } else {
+            Log.i("SCHEDULED_TASK_RUN", "Wallet backup DOES NOT exist")
+            false
+        }
+    }
+
     @ReactMethod
     fun walletExists(promise: Promise) {
         // Check if a wallet already exists
@@ -381,9 +393,14 @@ class RPCModule internal constructor(private val reactContext: ReactApplicationC
 
     @ReactMethod
     fun deleteExistingWallet(promise: Promise) {
-        val file = MainApplication.getAppContext()?.getFileStreamPath(walletFileName)
-        if (file!!.delete()) {
-            promise.resolve(true)
+        // check first if the file exists
+        if (wallet_exists()) {
+            val file = MainApplication.getAppContext()?.getFileStreamPath(walletFileName)
+            if (file!!.delete()) {
+                promise.resolve(true)
+            } else {
+                promise.resolve(false)
+            }
         } else {
             promise.resolve(false)
         }
@@ -391,9 +408,14 @@ class RPCModule internal constructor(private val reactContext: ReactApplicationC
 
     @ReactMethod
     fun deleteExistingWalletBackup(promise: Promise) {
-        val file = MainApplication.getAppContext()?.getFileStreamPath(walletBackupFileName)
-        if (file!!.delete()) {
-            promise.resolve(true)
+        // check first if the file exists
+        if (walletBackup_exists()) {
+            val file = MainApplication.getAppContext()?.getFileStreamPath(walletBackupFileName)
+            if (file!!.delete()) {
+                promise.resolve(true)
+            } else {
+                promise.resolve(false)
+            }
         } else {
             promise.resolve(false)
         }
